// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  ai: typeof ai;
  arrivages: typeof arrivages;
  articles: typeof articles;
  ateliers: typeof ateliers;
  bennespro: typeof bennespro;
  bennesproClientVehicles: typeof bennesproClientVehicles;
  bikes: typeof bikes;
  clerkMigration: typeof clerkMigration;
  clients: typeof clients;
  community: typeof community;
  crons: typeof crons;
  dashboard: typeof dashboard;
  documents: typeof documents;
  emails: typeof emails;
  files: typeof files;
  fleet: typeof fleet;
  gotravaux: typeof gotravaux;
  http: typeof http;
  importLegacy: typeof importLegacy;
  klyde: typeof klyde;
  leaves: typeof leaves;
  lib: typeof lib;
  livraison: typeof livraison;
  mesoutilsEmails: typeof mesoutilsEmails;
  mesoutilsNotifications: typeof mesoutilsNotifications;
  messages: typeof messages;
  notifications: typeof notifications;
  permissions: typeof permissions;
  pointeuse: typeof pointeuse;
  posts: typeof posts;
  processes: typeof processes;
  requestAnalysis: typeof requestAnalysis;
  requests: typeof requests;
  reservations: typeof reservations;
  sorties: typeof sorties;
  stripe: typeof stripe;
  team: typeof team;
  users: typeof users;
  vehicleControlReminders: typeof vehicleControlReminders;
  ventes: typeof ventes;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;